+++ packages/api/src/resolvers/queries/index.ts
@@ -65,6 +65,8 @@ import token from './warehousing/token.js';
 import work from './worker/work.js';
 import workQueue from './worker/workQueue.js';
 import workQueueCount from './worker/workQueueCount.js';
+import validateResetPasswordToken from './users/validateResetPasswordToken.js';
+import validateVerifyEmailToken from './users/validateVerifyEmailToken.js';
 
 export default {
   me,
@@ -132,4 +134,6 @@ export default {
   event: acl(actions.viewEvent)(event),
   events: acl(actions.viewEvents)(events),
   eventsCount: acl(actions.viewEvents)(eventsCount),
+  validateResetPasswordToken: acl(actions.resetPassword)(validateResetPasswordToken),
+  validateVerifyEmailToken: acl(actions.verifyEmail)(validateVerifyEmailToken),
 };