+++ packages/api/src/schema/query.ts
@@ -455,6 +455,16 @@ export default [
       Get total count of all emitted events
       """
       eventsCount(types: [String!], queryString: String, created: DateTime): Int!
+
+      """
+      Determines if a token is valid/active for reset password
+      """
+      validateResetPasswordToken(token: String!): Boolean!
+
+      """
+      Determines if a token is valid/active for email verification
+      """
+      validateVerifyEmailToken(token: String!): Boolean!
     }
   `,
 ];