+++ changelog.md
@@ -1,4 +1,6 @@
 # Unchained Engine vNEXT
+- Add `Query.validateVerifyEmailToken` that can be used to verify if a token is valid for use when verifying email
+- Add `Query.validateResetPasswordToken` that can be used to verify if a token is valid for use on password reset request
 
 ## Removing the auth fat of unchained
 We experienced feature creep in the authentication part of Unchained and suddenly woke up to homemade implementations of Two-Factor Auth via TOTP, WebAuthn, oAuth, Impersonator features etc. Many solutions like Zitadel, Keycloak, Auth0 etc. solve that just perfect and keep up with the ever increasing complexity of auth mechanisms. At the same time, core-accountsjs depends on a package called accountsjs which is unmaintained and uses a conflicting old mongodb driver.